backend/pyproject.toml
[tool.poetry]
name = "student-evaluation-system-backend"
version = "0.1.0"
description = "UCL student application evaluation system backend"
authors = ["UCL Eval Team <dev@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
pydantic-settings = "^2.3.4"
SQLAlchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
python-multipart = "^0.0.9"
celery = "^5.4.0"
redis = "^5.0.7"
beautifulsoup4 = "^4.12.3"
semantic-kernel = "^1.16.1"
azure-identity = "^1.18.0"
azure-ai-agents = "^1.0.0b2"
azure-ai-documentintelligence = "^1.0.0b5"
httpx = "^0.27.2"
tenacity = "^9.0.0"
pycountry = "^24.6.1"
pandas = "^2.2.0"
lxml = "^5.2.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.8"
mypy = "^1.11.2"
types-redis = "^4.6.0.20240819"
types-beautifulsoup4 = "^4.12.0.20240822"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
